+++ src/types/habits.ts
@@ -1,0 +1,10 @@
+export interface DayEntry {
+  workout: number;
+  eating: number;
+  reading: number;
+  sleep: number;
+}
+
+export interface HabitData {
+  [date: string]: DayEntry;
+}
